feat: detect allows dask tests: rename pipeline to e2e
